Completed the get answer method.
Why:To identify the type of question and according to the answer, we
have to show the child question and make failled statement according to
the answer.

How:By getting the answer object and type of answer in question,
according to question type, we may pass or fail.

Tags:PM question answer form.